requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiolimiter==1.1.0
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.1.0
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.2
executing==2.0.1
faiss-cpu==1.13.0
fastapi==0.112.0
filelock==3.20.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2025.10.0
genai==2.1.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
hf-xet==1.2.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.36.0
ibm-cloud-sdk-core==3.18.0
ibm-cos-sdk==2.13.6
ibm-cos-sdk-core==2.13.6
ibm-cos-sdk-s3transfer==2.13.6
ibm-generative-ai==0.1.17
ibm-watson==8.1.0
ibm-watson-machine-learning==1.0.335
ibm_watsonx_ai==1.1.2
idna==3.7
ImageIO==2.37.2
importlib-metadata==6.11.0
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.2.14
langchain-cli==0.0.29
langchain-community==0.2.11
langchain-core==0.2.34
langchain-experimental==0.0.64
langchain-ibm==0.1.12
langchain-openai==0.1.20
langchain-text-splitters==0.2.2
langserve==0.2.2
langsmith==0.1.96
libcst==1.4.0
lomond==0.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.8.2
matplotlib-inline==0.1.7
mdurl==0.1.2
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.4.2
numpy==1.26.4
openai==1.40.2
orjson==3.10.6
packaging==23.2
pandas==1.5.3
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
prompt_toolkit==3.0.47
protobuf==4.25.4
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
PyAutoGUI==0.9.54
pydantic==1.10.10
pydantic_core==2.20.1
pydeck==0.9.1
PyGetWindow==0.0.9
Pygments==2.18.0
PyJWT==2.9.0
PyMsgBox==2.0.1
pyobjc-core==12.1
pyobjc-framework-Cocoa==12.1
pyobjc-framework-Quartz==12.1
pyparsing==3.1.2
pyperclip==1.11.0
pyproject-toml==0.0.10
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytweening==1.2.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.7.24
requests==2.32.2
rich==13.7.1
rpds-py==0.19.1
rubicon-objc==0.5.2
safetensors==0.7.0
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
sentence-transformers==5.1.2
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.31
sse-starlette==1.8.2
stack-data==0.6.3
starlette==0.37.2
streamlit==1.29.0
streamlit-chat==0.1.1
sympy==1.14.0
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.22.1
toml==0.10.2
tomlkit==0.12.5
toolz==0.12.1
torch==2.9.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.57.1
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.19
uvicorn==0.23.2
validators==0.33.0
wcwidth==0.2.13
websocket-client==1.8.0
wxai-langchain==0.0.4
yarl==1.9.4
zipp==3.19.2
